Several small 16 vs 32 bit casts
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@5056 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/asn.1/asn1_encode.c
@@ -38,7 +38,7 @@ asn1_error_code INTERFACE asn1_encode_integer(buf, val, retlen)
   
   valcopy = val;
   do {
-    digit = valcopy&0xFF;
+    digit = (int) (valcopy&0xFF);
     retval = asn1buf_insert_octet(buf,(asn1_octet) digit);
     if(retval) return retval;
     length++;
@@ -75,7 +75,7 @@ asn1_error_code INTERFACE asn1_encode_unsigned_integer(buf, val, retlen)
   
   valcopy = val;
   do {
-    digit = valcopy&0xFF;
+    digit = (int) (valcopy&0xFF);
     retval = asn1buf_insert_octet(buf,(asn1_octet) digit);
     if(retval) return retval;
     length++;
